+++ src/main/java/insane96mcp/insanelib/InsaneLib.java
@@ -6,7 +6,6 @@ import com.google.gson.TypeAdapterFactory;
 import com.mojang.logging.LogUtils;
 import insane96mcp.insanelib.data.AttributeModifierOperationSerializer;
 import insane96mcp.insanelib.data.JsonFeatureDataReloadListener;
-import insane96mcp.insanelib.module.Modules;
 import insane96mcp.insanelib.network.NetworkHandler;
 import insane96mcp.insanelib.setup.ILModConfig;
 import insane96mcp.insanelib.util.IntegratedPack;
@@ -29,8 +28,8 @@ public class InsaneLib {
     public static ILModConfig CONFIG;
 
     public InsaneLib(IEventBus modEventBus, ModContainer modContainer) {
-        CONFIG = new ILModConfig(MOD_ID, ModConfig.Type.COMMON, modEventBus,
-                Modules::init, InsaneLib.class.getClassLoader());
+        CONFIG = new ILModConfig(location("base"), "Base", ModConfig.Type.COMMON,
+                modEventBus, InsaneLib.class.getClassLoader());
         modContainer.registerConfig(ModConfig.Type.COMMON, CONFIG.spec, MOD_ID + "/common.toml");
         modEventBus.addListener(IntegratedPack::onAddPackFinders);
         modEventBus.addListener(NetworkHandler::register);